+++ NEWS.md
@@ -2,6 +2,8 @@
 
 ## Changes in 1.7.3
 
+- Add `SingleMatchParam` for `filterMatches` to allow selection of (at most) a
+  single match to a target element for each query element.
 - Add new methods `queryVariables` and `targetVariables` to extract the names
   of variables (columns) of *query* and *target*.